+++ packages/moonstone/GridListImageItem/GridListImageItem.js
@@ -13,7 +13,7 @@ import Spottable from '@enact/spotlight/Spottable';
 
 import Icon from '../Icon';
 import {Image} from '../Image';
-import {MarqueeController, MarqueeText} from '../Marquee';
+import {Marquee, MarqueeController} from '../Marquee';
 import Skinnable from '../Skinnable';
 
 import css from './GridListImageItem.less';
@@ -163,8 +163,8 @@ const GridListImageItemBase = kind({
 				<Image className={css.image} placeholder={placeholder} src={source}>
 					{selectionOverlay}
 				</Image>
-				{caption ? (<MarqueeText alignment="center" className={css.caption} marqueeOn="hover">{caption}</MarqueeText>) : null}
-				{subCaption ? (<MarqueeText alignment="center" className={css.subCaption} marqueeOn="hover">{subCaption}</MarqueeText>) : null}
+				{caption ? (<Marquee alignment="center" className={css.caption} marqueeOn="hover">{caption}</Marquee>) : null}
+				{subCaption ? (<Marquee alignment="center" className={css.subCaption} marqueeOn="hover">{subCaption}</Marquee>) : null}
 			</div>
 		);
 	}